# Makes spline code safer and adds a more thorough test case