--- a/src/math/point.rs
+++ b/src/math/point.rs
@@ -1,8 +1,10 @@
+use std::ops::{Add, Div, Mul, Sub};
+
 use num::{cast, Float, NumCast};
 
 /// Represents a 2D point (or any pair of objects).
 #[allow(missing_docs)]
-#[derive(Add, Sub, Mul, Div, Clone, Copy, Default, Debug, Display, PartialEq, Eq, Hash)]
+#[derive(Clone, Copy, Default, Debug, Display, PartialEq, Eq, Hash)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[display(fmt = "({}, {})", "x", "y")]
 pub struct Point<T> {
@@ -14,6 +16,50 @@ impl<T> Point<T> {
     /// Create a new point
     pub fn new(x: T, y: T) -> Point<T> {
         Point { x, y }
+    }
+}
+
+impl<T: Div<Output = T>> Div for Point<T> {
+    type Output = Self;
+
+    fn div(self, rhs: Self) -> Self::Output {
+        Self {
+            x: self.x / rhs.x,
+            y: self.y / rhs.y,
+        }
+    }
+}
+
+impl<T: Mul<Output = T>> Mul for Point<T> {
+    type Output = Self;
+
+    fn mul(self, rhs: Self) -> Self::Output {
+        Self {
+            x: self.x * rhs.x,
+            y: self.y * rhs.y,
+        }
+    }
+}
+
+impl<T: Add<Output = T>> Add for Point<T> {
+    type Output = Self;
+
+    fn add(self, rhs: Self) -> Self::Output {
+        Self {
+            x: self.x + rhs.x,
+            y: self.y + rhs.y,
+        }
+    }
+}
+
+impl<T: Sub<Output = T>> Sub for Point<T> {
+    type Output = Self;
+
+    fn sub(self, rhs: Self) -> Self::Output {
+        Self {
+            x: self.x - rhs.x,
+            y: self.y - rhs.y,
+        }
     }
 }
 
@@ -29,9 +75,27 @@ impl<T: Float> Point<T> {
     /// Calculates the Euclidean distance between 2 points.
     #[inline]
     pub fn distance(&self, other: Point<T>) -> T {
+        self.distance_squared(other).sqrt()
+    }
+
+    /// Calculates the Euclidean distance squared between 2 points. Faster
+    #[inline]
+    pub fn distance_squared(&self, other: Point<T>) -> T {
         let dx = other.x.sub(self.x);
         let dy = other.y.sub(self.y);
-        (dx * dx + dy * dy).sqrt()
+        dx * dx + dy * dy
+    }
+
+    /// Calculates the Euclidean distance between this point and origin.
+    #[inline]
+    pub fn length(&self) -> T {
+        self.length_squared().sqrt()
+    }
+
+    /// Calculates the Euclidean distance squared between this point and origin. Faster
+    #[inline]
+    pub fn length_squared(&self) -> T {
+        self.x * self.x + self.y * self.y
     }
 
     /// Calculates the magnitude of the vector.