… tests/spline_points_list.in