Also seems to improve performance slightly for some reason
Before at 10'000 iterations:
Debug - 538ms
Release - 32ms

After at 10'000 iterations:
Debug - 465ms
Release - 29ms