@nobbele 2021-01-31
FYI the test code uses the same coordinates returned before this PR so the result is identical